move ordered preference list methods in a helper class
